+++ upcloud/resource_upcloud_managed_object_storage_user_test.go
@@ -36,8 +36,32 @@ func TestAccUpcloudManagedObjectStorageUser(t *testing.T) {
 				),
 			},
 			{
-				Config:            testDataS2,
+				Config:            testDataS1,
+				ResourceName:      user,
+				ImportState:       true,
 				ImportStateVerify: true,
+			},
+			{
+				Config:            testDataS1,
+				ResourceName:      userPolicy,
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
+			{
+				Config:            testDataS1,
+				ResourceName:      policy,
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
+			{
+				Config:                  testDataS1,
+				ResourceName:            userAccessKey,
+				ImportState:             true,
+				ImportStateVerify:       true,
+				ImportStateVerifyIgnore: []string{"secret_access_key"}, // only provided on creation, not available on subsequent requests like import
+			},
+			{
+				Config: testDataS2,
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr(storage, "name", "tf-acc-test-objstov2-user"),
 					resource.TestCheckResourceAttr(policy, "name", "tf-acc-test-objstov2-user"),